fix: boardRepository bean 생성 에러 해결
